Add noddyverse dataset
This adds a new dataset and wrapper that fits the existing training regime. The noddyverse dataset can be considered as a raster loader with custom resize function that grids a point sampled geophysical raster as the "downsampling" method.

It also requires adjusting the allowed value range - noddyverse is currently clipped to +-5000 then min-max normalised to [-1, 1].

Additionally, we change the channel count in some logic.