Fix training sample scale factors
We are training arbitrary scale factor SR. This means training on a variety of samples.
Since we are comparing to gridding methods that use all the line data available, we want to keep all the line data for our network.

The original LTE paper uses the sample_q parameter to preserve HR dimensions in a batch, and a fixed LR size. The scale factor comes into the scale that HR is vs LR (e.g. lr 48, scale 2, hr 96.

We will train at 4x line spacing for the high resolution grid (ostensibly the "groundtruth" of field data). This is because our default cell size to line spacing ratio is 1/4, meaning our resulting HR grid will be (approx) the same size as the ground truth noddy models. I say approx because we currently clip (d) to 180, which avoids some gridding artefacts.

To match the input size of LTE, we crop the HR and LR grids, at the configured size (and scaled size).

The scale factor is set randomly per batch in the configured range, so the tensors are consistent sizes and can be batched by pytorch.