Change dataset handling
Using a full (or at least, random selection) of the Noddyverse is critical for generalisation across geological terranes. Previously for development we used a subset split by folders. Now we use a (pre-shuffled) csv list of possible history files (and their relative paths), and a second list of "reserved" models (from the Noddy10k dataset). These models are randomly selected models from the 1M dataset. Care is taken to ensure no 10k test samples are present in train or val.

Additionally, train and val split is done by configuring in slice notation the start and end points in the dataset. Care must be taken to not overlap the two ranges in the config text file.