Add sampling noise configuration
Can now specify 2 types of noise to add to the synthetic sampling during inference (and probably training if you add the config to the dataset in train config).

Levelling noise is for line levelling issues, and is a random constant offset per line.
Gaussian noise is for sensor accuracy/noise issues and applied randomly to each pixel.

In each instance, Numpy RNG Normal is used, with scale 0.25 to generate a curve with min/max around 1. This scales the dynamic range of the sample, which itself is scaled by the decimal percentage provided in the config. The outcome is random noise at the configured % of the local TMI range, as per Holden et al. 2011.

Regional Geology noise as per the above paper is not included, because that noise is to simulate the effect of regional geology - Noddyverse already includes that.

Holden, E.-J., Fu, S.C., Kovesi, P., Dentith, M., Bourne, B., Hope, M., 2011. Automatic identification of responses from porphyry intrusive systems within magnetic data using image analysis. Journal of Applied Geophysics 74, 255–262. https://doi.org/10.1016/j.jappgeo.2011.06.016